feat(float): Added float data type for more realistic measurement

+++ WaterLogger.Ryanw84/Models/DrinkingWaterModel.cs
@@ -9,7 +9,7 @@ namespace WaterLogger.Ryanw84.Models
         [DisplayFormat(DataFormatString = "{0:dd-MM-yy}", ApplyFormatInEditMode = true)]
         public DateTime Date { get; set; }
 
-        [Range(0, Int32.MaxValue, ErrorMessage = "Value for {0} must be a positive.")]
-        public int Quantity { get; set; }
+        [Range(0, float.MaxValue, ErrorMessage = "Value for {0} must be a positive.")]
+        public float Quantity { get; set; }
     }
 }
